delete item respects user permissions

+++ backend/src/resolvers/Mutation.js
@@ -46,10 +46,17 @@ const Mutations = {
     const where = { id: args.id };
 
     // find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
 
-    // check if they have permissions
-    // TODO
+    // check if user owns item, or if they have permissions
+    const ownsItem = item.user.id == ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => (
+      ['ADMIN','ITEMDELETE'].includes(permission)
+    ));
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have the required permissions to perform this action.")
+    }
 
     // delete item
     return ctx.db.mutation.deleteItem({ where }, info);
